+++ controllers/match_controller.go
@@ -49,6 +49,12 @@ func ReMatch(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
+
+	// Reverse order of players for rematch
+	for i := len(match.Players)/2 - 1; i >= 0; i-- {
+		opp := len(match.Players) - 1 - i
+		match.Players[i], match.Players[opp] = match.Players[opp], match.Players[i]
+	}
 	match.CreatedAt = time.Now().Format("2006-01-02 15:04:05")
 	match, err = data.NewMatch(*match)
 	if err != nil {